# groSCery-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "201 Project API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/groups/subscribe/{group_id}": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Subscribe a user to a group",
                "operationId": "App\\Http\\Controllers\\GroupController::subscribe",
                "responses": {
                    "200": {
                        "description": "User has been subscribed to group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "User has been subscribed to group 201"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/groups/unsubscribe": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Unsubscribe a user from a group",
                "operationId": "App\\Http\\Controllers\\GroupController::unsubscribe",
                "responses": {
                    "200": {
                        "description": "User has been unsubscribed from their group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "User has been unsubscribed to group 201"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/groups/subscribers": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Gets the list of users subscribed to users Group",
                "operationId": "App\\Http\\Controllers\\GroupController::subscribers",
                "responses": {
                    "200": {
                        "description": "Group subscribers returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": [
                                        {
                                            "id": 2,
                                            "group_id": null,
                                            "name": "Kristof",
                                            "pic_url": null,
                                            "email": "kristof@usc.edu",
                                            "created_at": "2018-10-30 01:13:02"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
...
        "/api/items/create": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Create a new item and subscribe to it",
                "operationId": "App\\Http\\Controllers\\ItemController::create",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item created and user subscribed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": {
                                        "id": "1",
                                        "group_id": "1",
                                        "name": "Banana",
                                        "in_stock": "1"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Item created and user subscribed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "Item already exists"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/items/subscribers/{item_id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Gets the list of users subscribed to a Item",
                "operationId": "App\\Http\\Controllers\\ItemController::subscribers",
                "responses": {
                    "200": {
                        "description": "Item subscribers returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": [
                                        {
                                            "id": 2,
                                            "group_id": null,
                                            "name": "Kristof",
                                            "pic_url": null,
                                            "email": "kristof@usc.edu",
                                            "created_at": "2018-10-30 01:13:02"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/items/subscribe/{item_id}": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Subscribe a user to an Item",
                "operationId": "App\\Http\\Controllers\\ItemController::subscribe",
                "responses": {
                    "200": {
                        "description": "User subscribed to Item",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": {
                                        "id": "1",
                                        "group_id": "1",
                                        "name": "Banana",
                                        "in_stock": "1"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/items/unsubscribe/{item_id}": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Unsubscribe a user to an Item",
                "operationId": "App\\Http\\Controllers\\ItemController::unsubscribe",
                "responses": {
                    "200": {
                        "description": "User unsubscribed to Item",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": {
                                        "id": "1",
                                        "group_id": "1",
                                        "name": "Banana",
                                        "in_stock": "1"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/items/in-stock/{item_id}": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Mark a item as in stock",
                "operationId": "App\\Http\\Controllers\\ItemController::inStock",
                "responses": {
                    "200": {
                        "description": "Item indicated as in stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "Item marked in stock"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "Item not found"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/items/out-of-stock/{item_id}": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Mark a item as out of stock",
                "operationId": "App\\Http\\Controllers\\ItemController::outOfStock",
                "responses": {
                    "200": {
                        "description": "Item indicated as out of stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "Item marked out of stock"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "Item not found"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
...
    },
    "components": {
        "responses": {
            "401": {
                "description": "Request is not authorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "error": "Unauthorized"
                        }
                    }
                }
            },
            "400": {
                "description": "Bad request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "error": "Bad request"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT"
            }
        }
    }
}